requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.30
pymysql==1.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.3.0
